Refactor into single file and changed config names
