--- src/zeroband/inference/pipeline/__init__.py
@@ -1,4 +1,0 @@
-from .hooks import setup_hooks
-from .comm import setup_comm
-
-__all__ = ["setup_hooks", "setup_comm"]
